requirements.txt
certifi==2024.12.14; python_version >= '3.6'
charset-normalizer==3.4.1; python_version >= '3.7'
dnspython==2.7.0; python_version >= '3.9'
idna==3.10; python_version >= '3.6'
pymongo==4.10.1; python_version >= '3.8'
python-dotenv==1.0.1; python_version >= '3.8'
requests==2.32.3; python_version >= '3.8'
urllib3==2.3.0; python_version >= '3.9'
